nav up and down back in notePage
