+++ src/components/NotePage/NotePage.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo, useCallback } from 'react'
+import React, { useMemo, useCallback, useState } from 'react'
 import NoteList from './NoteList'
 import styled from '../../lib/styled'
 import NoteDetail from './NoteDetail'
@@ -8,7 +8,8 @@ import {
   StorageNotesRouteParams,
   StorageTrashCanRouteParams,
   StorageTagsRouteParams,
-  usePathnameWithoutNoteId
+  usePathnameWithoutNoteId,
+  useRouter
 } from '../../lib/router'
 import { useDb } from '../../lib/db'
 import TwoPaneLayout from '../atoms/TwoPaneLayout'
@@ -40,7 +41,8 @@ export default () => {
     if (storageId == null) return undefined
     return db.storageMap[storageId]
   }, [db.storageMap, storageId])
-
+  const router = useRouter()
+  const [search, setSearchInput] = useState<string>('')
   const currentPathnameWithoutNoteId = usePathnameWithoutNoteId()
 
   const notes = useMemo((): NoteDoc[] => {
@@ -77,18 +79,29 @@ export default () => {
     return []
   }, [currentStorage, routeParams])
 
+  const filteredNotes = useMemo(() => {
+    if (search.trim() === '') return notes
+    const regex = new RegExp(search, 'i')
+    return notes.filter(
+      note =>
+        note.tags.join().match(regex) ||
+        note.title.match(regex) ||
+        note.content.match(regex)
+    )
+  }, [search, notes])
+
   const currentNoteIndex = useMemo(() => {
-    for (let i = 0; i < notes.length; i++) {
-      if (notes[i]._id === noteId) {
+    for (let i = 0; i < filteredNotes.length; i++) {
+      if (filteredNotes[i]._id === noteId) {
         return i
       }
     }
     return 0
-  }, [notes, noteId])
+  }, [filteredNotes, noteId])
 
   const currentNote = useMemo(() => {
-    return notes[currentNoteIndex]
-  }, [notes, currentNoteIndex])
+    return filteredNotes[currentNoteIndex]
+  }, [filteredNotes, currentNoteIndex])
 
   const createNote = useCallback(async () => {
     if (storageId == null || routeParams.name === 'storages.trashCan') {
@@ -163,6 +176,24 @@ export default () => {
     [storageId, db]
   )
 
+  const navigateUp = useCallback(() => {
+    if (currentNoteIndex > 0) {
+      router.push(
+        currentPathnameWithoutNoteId +
+          `/${filteredNotes[currentNoteIndex - 1]._id}`
+      )
+    }
+  }, [filteredNotes, currentNoteIndex, router, currentPathnameWithoutNoteId])
+
+  const navigateDown = useCallback(() => {
+    if (currentNoteIndex < notes.length - 1) {
+      router.push(
+        currentPathnameWithoutNoteId +
+          `/${filteredNotes[currentNoteIndex + 1]._id}`
+      )
+    }
+  }, [filteredNotes, currentNoteIndex, router, currentPathnameWithoutNoteId])
+
   return storageId != null ? (
     <TwoPaneLayout
       style={{ height: '100%' }}
@@ -170,12 +201,15 @@ export default () => {
       left={
         <FileDropZone style={{ height: '100%' }} onDrop={importDrop}>
           <NoteList
+            search={search}
+            setSearchInput={setSearchInput}
             storageId={storageId}
-            notes={notes}
+            notes={filteredNotes}
             createNote={createNote}
             basePathname={currentPathnameWithoutNoteId}
-            currentPathnameWithoutNoteId={currentPathnameWithoutNoteId}
-            noteId={noteId}
+            navigateDown={navigateDown}
+            navigateUp={navigateUp}
+            currentNoteIndex={currentNoteIndex}
           />
         </FileDropZone>
       }
